refactor: `feignClient` `common` 도메인으로 이동
