+++ src/main/java/com/playus/twpservice/domain/chat/service/ChatRoomService.java
@@ -16,8 +16,8 @@ import com.playus.twpservice.domain.chat.repository.write.ChatRoomRepository;
 import com.playus.twpservice.domain.chat.repository.message.custom.ChatMessageRepositoryCustom;
 import com.playus.twpservice.domain.party.entity.Party;
 import com.playus.twpservice.domain.party.exception.entity.PartyException;
-import com.playus.twpservice.domain.party.feign.client.UserFeignClient;
-import com.playus.twpservice.domain.party.feign.response.UserInfoResponse;
+import com.playus.twpservice.domain.common.feign.client.UserFeignClient;
+import com.playus.twpservice.domain.common.feign.response.UserInfoResponse;
 import com.playus.twpservice.domain.chat.kafka.KafkaChatPublisher;
 import com.playus.twpservice.domain.party.repository.write.PartyRepository;
 import lombok.RequiredArgsConstructor;
