+++ src/main/java/com/playus/twpservice/domain/party/service/PartyService.java
@@ -34,8 +34,8 @@ import com.playus.twpservice.domain.party.enums.PartyAgeGroup;
 import com.playus.twpservice.domain.party.enums.PartyJoinMethod;
 import com.playus.twpservice.domain.party.enums.PartyJoinRequestStatus;
 import com.playus.twpservice.domain.party.exception.entity.PartyException;
-import com.playus.twpservice.domain.party.feign.client.NotificationFeignClient;
-import com.playus.twpservice.domain.party.feign.event.PartyNotificationEvent;
+import com.playus.twpservice.domain.common.feign.client.NotificationFeignClient;
+import com.playus.twpservice.domain.common.feign.event.PartyNotificationEvent;
 import com.playus.twpservice.domain.party.repository.read.PartyAgeReadOnlyRepository;
 import com.playus.twpservice.domain.party.repository.read.PartyJoinReadOnlyRepository;
 import com.playus.twpservice.domain.party.repository.read.PartyReadOnlyRepository;
