+++ src/test/java/com/playus/twpservice/domain/party/feign/client/MatchFeignClientTest.java
@@ -1,6 +1,7 @@
 package com.playus.twpservice.domain.party.feign.client;
 
 import com.playus.twpservice.IntegrationTestSupport;
+import com.playus.twpservice.domain.common.feign.client.MatchFeignClient;
 import com.playus.twpservice.global.response.ErrorResponse;
 import feign.FeignException;
 import org.junit.jupiter.api.DisplayName;
